+++ src/main/java/net/frozenblock/lib/worldgen/feature/api/FrozenLibFeatures.java
@@ -24,7 +24,6 @@ import net.frozenblock.lib.worldgen.feature.api.features.CircularWaterloggedVege
 import net.frozenblock.lib.worldgen.feature.api.features.CircularWaterloggedVegetationPatchLessBordersFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.ColumnWithDiskFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.ComboFeature;
-import net.frozenblock.lib.worldgen.feature.api.features.CurvingTunnelFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.DownwardsColumnFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.FadingDiskCarpetFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.FadingDiskFeature;
@@ -32,8 +31,8 @@ import net.frozenblock.lib.worldgen.feature.api.features.FadingDiskExceptInBiome
 import net.frozenblock.lib.worldgen.feature.api.features.FadingDiskScheduleTickFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.NoisePathFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.NoisePathScheduleTickFeature;
-import net.frozenblock.lib.worldgen.feature.api.features.NoisePathSwapUnderWaterFeature;
-import net.frozenblock.lib.worldgen.feature.api.features.NoisePathUnderWaterFeature;
+import net.frozenblock.lib.worldgen.feature.api.features.NoisePathSwapUnderFluidFeature;
+import net.frozenblock.lib.worldgen.feature.api.features.NoisePathNEARWaterFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.RequiresAirOrWaterInAreaNoisePathFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.SimpleBlockScheduleTickFeature;
 import net.frozenblock.lib.worldgen.feature.api.features.UpwardsColumnFeature;
@@ -41,12 +40,11 @@ import net.frozenblock.lib.worldgen.feature.api.features.config.AirOrWaterInArea
 import net.frozenblock.lib.worldgen.feature.api.features.config.ColumnFeatureConfig;
 import net.frozenblock.lib.worldgen.feature.api.features.config.ColumnWithDiskFeatureConfig;
 import net.frozenblock.lib.worldgen.feature.api.features.config.ComboFeatureConfig;
-import net.frozenblock.lib.worldgen.feature.api.features.config.CurvingTunnelFeatureConfig;
 import net.frozenblock.lib.worldgen.feature.api.features.config.FadingDiskCarpetFeatureConfig;
 import net.frozenblock.lib.worldgen.feature.api.features.config.FadingDiskFeatureConfig;
 import net.frozenblock.lib.worldgen.feature.api.features.config.FadingDiskWithBiomeFeatureConfig;
 import net.frozenblock.lib.worldgen.feature.api.features.config.PathFeatureConfig;
-import net.frozenblock.lib.worldgen.feature.api.features.config.PathSwapUnderWaterFeatureConfig;
+import net.frozenblock.lib.worldgen.feature.api.features.config.PathSwapUnderFluidFeatureConfig;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.levelgen.feature.configurations.SimpleBlockConfiguration;
@@ -55,8 +53,8 @@ import net.minecraft.world.level.levelgen.feature.configurations.VegetationPatch
 public class FrozenLibFeatures {
 	public static final RequiresAirOrWaterInAreaNoisePathFeature REQUIRES_AIR_OR_WATER_IN_AREA_NOISE_PATH_FEATURE = new RequiresAirOrWaterInAreaNoisePathFeature(AirOrWaterInAreaPathFeatureConfig.CODEC);
 	public static final NoisePathFeature NOISE_PATH_FEATURE = new NoisePathFeature(PathFeatureConfig.CODEC);
-	public static final NoisePathSwapUnderWaterFeature NOISE_PATH_SWAP_UNDER_WATER_FEATURE = new NoisePathSwapUnderWaterFeature(PathSwapUnderWaterFeatureConfig.CODEC);
-	public static final NoisePathUnderWaterFeature NOISE_PATH_UNDER_WATER_FEATURE = new NoisePathUnderWaterFeature(PathFeatureConfig.CODEC);
+	public static final NoisePathSwapUnderFluidFeature NOISE_PATH_SWAP_UNDER_FLUID_FEATURE = new NoisePathSwapUnderFluidFeature(PathSwapUnderFluidFeatureConfig.CODEC);
+	public static final NoisePathNEARWaterFeature NOISE_PATH_NEAR_WATER_FEATURE = new NoisePathNEARWaterFeature(PathFeatureConfig.CODEC);
 	public static final ColumnWithDiskFeature COLUMN_WITH_DISK_FEATURE = new ColumnWithDiskFeature(ColumnWithDiskFeatureConfig.CODEC);
 	public static final UpwardsColumnFeature UPWARDS_COLUMN_FEATURE = new UpwardsColumnFeature(ColumnFeatureConfig.CODEC);
 	public static final DownwardsColumnFeature DOWNWARDS_COLUMN_FEATURE = new DownwardsColumnFeature(ColumnFeatureConfig.CODEC);
@@ -66,7 +64,6 @@ public class FrozenLibFeatures {
 	public static final FadingDiskFeature FADING_DISK_FEATURE = new FadingDiskFeature(FadingDiskFeatureConfig.CODEC);
 	public static final FadingDiskScheduleTickFeature FADING_DISK_SCHEDULE_TICK_FEATURE = new FadingDiskScheduleTickFeature(FadingDiskFeatureConfig.CODEC);
 	public static final FadingDiskCarpetFeature FADING_DISK_CARPET_FEATURE = new FadingDiskCarpetFeature(FadingDiskCarpetFeatureConfig.CODEC);
-	public static final CurvingTunnelFeature CURVING_TUNNEL_FEATURE = new CurvingTunnelFeature(CurvingTunnelFeatureConfig.CODEC);
 	public static final CircularLavaVegetationPatchFeature CIRCULAR_LAVA_VEGETATION_PATCH = new CircularLavaVegetationPatchFeature(VegetationPatchConfiguration.CODEC);
 	public static final CircularLavaVegetationPatchLessBordersFeature CIRCULAR_LAVA_VEGETATION_PATCH_LESS_BORDERS = new CircularLavaVegetationPatchLessBordersFeature(VegetationPatchConfiguration.CODEC);
 	public static final SimpleBlockScheduleTickFeature SIMPLE_BLOCK_SCHEDULE_TICK_FEATURE = new SimpleBlockScheduleTickFeature(SimpleBlockConfiguration.CODEC);
@@ -76,8 +73,8 @@ public class FrozenLibFeatures {
 	public static void init() {
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("requires_air_or_water_in_area_noise_path_feature"), REQUIRES_AIR_OR_WATER_IN_AREA_NOISE_PATH_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("noise_path_feature"), NOISE_PATH_FEATURE);
-		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("noise_path_swap_under_water_feature"), NOISE_PATH_SWAP_UNDER_WATER_FEATURE);;
-		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("noise_path_under_water_feature"), NOISE_PATH_UNDER_WATER_FEATURE);
+		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("noise_path_swap_under_fluid_feature"), NOISE_PATH_SWAP_UNDER_FLUID_FEATURE);;
+		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("noise_path_near_water_feature"), NOISE_PATH_NEAR_WATER_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("column_with_disk_feature"), COLUMN_WITH_DISK_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("upwards_column"), UPWARDS_COLUMN_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("downwards_column"), DOWNWARDS_COLUMN_FEATURE);
@@ -87,7 +84,6 @@ public class FrozenLibFeatures {
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("fading_disk_feature"), FADING_DISK_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("fading_disk_schedule_tick_feature"), FADING_DISK_SCHEDULE_TICK_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("fading_disk_carpet_feature"), FADING_DISK_CARPET_FEATURE);
-		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("curving_tunnel_feature"), CURVING_TUNNEL_FEATURE);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("circular_lava_vegetation_patch"), CIRCULAR_LAVA_VEGETATION_PATCH);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("circular_lava_vegetation_patch_less_borders"), CIRCULAR_LAVA_VEGETATION_PATCH_LESS_BORDERS);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("simple_block_schedule_tick"), SIMPLE_BLOCK_SCHEDULE_TICK_FEATURE);