+++ src/main/java/net/frozenblock/lib/worldgen/feature/api/features/NoisePathNEARWaterFeature.java
@@ -18,30 +18,23 @@
 package net.frozenblock.lib.worldgen.feature.api.features;
 
 import com.mojang.serialization.Codec;
-import net.frozenblock.lib.math.api.EasyNoiseSampler;
+import net.frozenblock.lib.worldgen.feature.api.FrozenLibFeatureUtils;
 import net.frozenblock.lib.worldgen.feature.api.features.config.PathFeatureConfig;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.WorldGenLevel;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.levelgen.Heightmap.Types;
-import net.minecraft.world.level.levelgen.blockpredicates.BlockPredicate;
-import net.minecraft.world.level.levelgen.feature.Feature;
-import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
-import net.minecraft.world.level.levelgen.synth.ImprovedNoise;
 import org.jetbrains.annotations.NotNull;
 
-public class NoisePathScheduleTickFeature extends NoisePathFeature {
+public class NoisePathNEARWaterFeature extends NoisePathFeature {
 
-	public NoisePathScheduleTickFeature(Codec<PathFeatureConfig> codec) {
+	public NoisePathNEARWaterFeature(Codec<PathFeatureConfig> codec) {
 		super(codec);
 	}
 
 	@Override
 	public boolean placeBlock(@NotNull WorldGenLevel level, BlockState state, BlockPos pos) {
-		super.placeBlock(level, state, pos);
-		level.scheduleTick(pos, state.getBlock(), 1);
-		return true;
+		if (!FrozenLibFeatureUtils.isWaterNearby(level, pos, 2)) return false;
+		return super.placeBlock(level, state, pos);
 	}
+
 }