+++ src/main/java/net/frozenblock/lib/worldgen/feature/api/features/RequiresAirOrWaterInAreaNoisePathFeature.java
@@ -19,13 +19,14 @@ package net.frozenblock.lib.worldgen.feature.api.features;
 
 import com.mojang.serialization.Codec;
 import net.frozenblock.lib.math.api.EasyNoiseSampler;
+import net.frozenblock.lib.worldgen.feature.api.FrozenLibFeatureUtils;
 import net.frozenblock.lib.worldgen.feature.api.features.config.AirOrWaterInAreaPathFeatureConfig;
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.WorldGenLevel;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.Heightmap.Types;
-import net.minecraft.world.level.levelgen.blockpredicates.BlockPredicate;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
 import net.minecraft.world.level.levelgen.synth.ImprovedNoise;
@@ -45,14 +46,8 @@ public class RequiresAirOrWaterInAreaNoisePathFeature extends Feature<AirOrWater
 		WorldGenLevel level = context.level();
 		int radiusSquared = config.radius() * config.radius();
 		RandomSource random = level.getRandom();
-		long noiseSeed = level.getSeed();
-		ImprovedNoise sampler =
-			config.noise() == 1 ? EasyNoiseSampler.createLocalNoise(noiseSeed) :
-				config.noise() == 2 ? EasyNoiseSampler.createCheckedNoise(noiseSeed) :
-					config.noise() == 3 ? EasyNoiseSampler.createLegacyThreadSafeNoise(noiseSeed) :
-						EasyNoiseSampler.createXoroNoise(noiseSeed);
+		ImprovedNoise sampler = config.noiseType().createNoise(level.getSeed());
 		float chance = config.placementChance();
-		int searchDistance = config.airSearchDistance();
 		int bx = blockPos.getX();
 		int by = blockPos.getY();
 		int bz = blockPos.getZ();
@@ -65,11 +60,7 @@ public class RequiresAirOrWaterInAreaNoisePathFeature extends Feature<AirOrWater
 					if (distance < radiusSquared) {
 						mutable.set(x, level.getHeight(Types.OCEAN_FLOOR, x, z) - 1, z);
 						double sample = EasyNoiseSampler.sample(sampler, mutable, config.noiseScale(), config.scaleY(), config.useY());
-						if (sample > config.minThreshold() && sample < config.maxThreshold() && level.getBlockState(mutable).is(config.replaceableBlocks())
-							&& checkSurroundingBlocks(level, mutable, searchDistance) && random.nextFloat() <= chance) {
-							generated = true;
-							level.setBlock(mutable, config.state().getState(random, mutable), Block.UPDATE_ALL);
-						}
+						generated = this.attemptPlaceBlock(config, level, random, chance, mutable, sample) || generated;
 					}
 				} else {
 					for (int y = by - config.radius(); y <= by + config.radius(); y++) {
@@ -77,11 +68,7 @@ public class RequiresAirOrWaterInAreaNoisePathFeature extends Feature<AirOrWater
 						if (distance < radiusSquared) {
 							mutable.set(x, y, z);
 							double sample = EasyNoiseSampler.sample(sampler, mutable, config.noiseScale(), config.scaleY(), config.useY());
-							if (sample > config.minThreshold() && sample < config.maxThreshold() && level.getBlockState(mutable).is(config.replaceableBlocks())
-								&& checkSurroundingBlocks(level, mutable, searchDistance) && random.nextFloat() <= chance) {
-								generated = true;
-								level.setBlock(mutable, config.state().getState(random, mutable), Block.UPDATE_ALL);
-							}
+							generated = this.attemptPlaceBlock(config, level, random, chance, mutable, sample) || generated;
 						}
 					}
 				}
@@ -90,17 +77,28 @@ public class RequiresAirOrWaterInAreaNoisePathFeature extends Feature<AirOrWater
 		return generated;
 	}
 
-	private static boolean checkSurroundingBlocks(WorldGenLevel level, @NotNull BlockPos pos, int searchDistance) {
-		Iterable<BlockPos> poses = BlockPos.betweenClosed(
-			pos.offset(-searchDistance, -searchDistance, -searchDistance),
-			pos.offset(searchDistance, searchDistance, searchDistance)
-		);
-		for (BlockPos blockPos : poses) {
-			if (BlockPredicate.ONLY_IN_AIR_OR_WATER_PREDICATE.test(level, blockPos)) {
-				return true;
-			}
+	public boolean attemptPlaceBlock(
+		@NotNull AirOrWaterInAreaPathFeatureConfig config,
+		WorldGenLevel level,
+		RandomSource random,
+		float chance,
+		BlockPos.MutableBlockPos mutable,
+		double sample
+	) {
+		if (sample > config.minThreshold()
+			&& sample < config.maxThreshold()
+			&& level.getBlockState(mutable).is(config.replaceableBlocks())
+			&& FrozenLibFeatureUtils.isAirOrWaterNearby(level, mutable, config.airOrWaterSearchDistance())
+			&& random.nextFloat() <= chance
+		) {
+			return this.placeBlock(level, config.state().getState(random, mutable), mutable);
 		}
 		return false;
+	}
+
+	public boolean placeBlock(@NotNull WorldGenLevel level, BlockState state, BlockPos pos) {
+		level.setBlock(pos, state, Block.UPDATE_CLIENTS);
+		return true;
 	}
 
 }