+++ src/main/java/net/frozenblock/lib/worldgen/feature/api/placementmodifier/LowerHeightmapPlacement.java
@@ -26,6 +26,8 @@ import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.placement.PlacementContext;
 import net.minecraft.world.level.levelgen.placement.PlacementModifier;
 import net.minecraft.world.level.levelgen.placement.PlacementModifierType;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
 
 public class LowerHeightmapPlacement extends PlacementModifier {
 	public static final MapCodec<LowerHeightmapPlacement> CODEC = RecordCodecBuilder.mapCodec(instance -> instance.group(
@@ -38,12 +40,13 @@ public class LowerHeightmapPlacement extends PlacementModifier {
 		this.heightmap = heightmap;
 	}
 
-	public static LowerHeightmapPlacement onHeightmap(Heightmap.Types heightmap) {
+	@Contract("_ -> new")
+	public static @NotNull LowerHeightmapPlacement onHeightmap(Heightmap.Types heightmap) {
 		return new LowerHeightmapPlacement(heightmap);
 	}
 
 	@Override
-	public Stream<BlockPos> getPositions(PlacementContext context, RandomSource random, BlockPos pos) {
+	public @NotNull Stream<BlockPos> getPositions(@NotNull PlacementContext context, RandomSource random, @NotNull BlockPos pos) {
 		int x = pos.getX();
 		int z = pos.getZ();
 		int y = context.getHeight(this.heightmap, x, z) - 1;